requirements.txt
PySide6
pillow
pywin32 
opencv-python
keyboard
pytesseract
pyautogui